Expanded Ground Scroll, updated explosions
I extended the ground scroll function to take over the functionality of
another script that was no longer necessary. I also expanded it to be
able to be applied to player and enemy bullets. I hope to make them
conform to the horizontal scrolling of the game without being hard to
predict.

I also updated the FX manager to instantiate and call more kinds of
effects with less coding required per new effect, which will help with
polish. I added a player muzzle flash as well.