More sounds, boss spawning fix, player game over fix
Fixed up the game manager to only tell the enemy spawner to spawn one
boss skull character. Fixed the player health to call game over properly
when the player reaches 0 health. I've almost implemented a state
machine for the player to make it easier to keep track of what it's
supposed to be doing, and I also implemented burst fire for the boss
character.