Added fade in/out for game start/end
I added code to the game manager and the UI manager to make them start
the game with a white overlay that fades out, and then fade the overlay
back in when the game is over. I have some class framework in the UI
manager that I may still find useful.